requirements.txt
vk_api
bs4
pyTelegramBotAPI
python-dotenv
m3u8